adds sonar.pluginKey property
This is need for a deployment to the SonarQube marketplace. The plugin key must be alphanumeric and is (by default) retrieved from the project's artifact id by stripping the leading "sonar-" and the trailing "-plugin". In our case the result "auth-oidc" is not compliant. So the plugin key is specified via the property "sonar.pluginKey" instead,